+++ CHANGELOG.md
@@ -2765,3 +2765,4 @@ Removed kcal from the macro donut chart entirely since macro-derived calories (p
 ## 2026-07-16
 - Made the empty state of Recent Foods card also show the Suggested card so users know it exists
 - Made primary and secondary muscles in the Todays Overview card show even when empty
+- Removed the set recent foods option in settings and removed all backend-related code to updating the recent foods max as it is now just handled based on premium status